Throw error is halo size is inappropriate for NonhydrostaticModel
